+++ saleor/core/middleware.py
@@ -1,4 +1,3 @@
-import datetime
 import logging
 from functools import wraps
 from typing import Callable
@@ -17,6 +16,7 @@ from django.conf import settings
 from django.contrib.sites.models import Site
 from django.core.exceptions import MiddlewareNotUsed
 from django.urls import reverse
+from django.utils import timezone
 from django.utils.functional import SimpleLazyObject
 from django.utils.translation import get_language
 from django_countries.fields import Country
@@ -106,9 +106,7 @@ def discounts(get_response):
     """Assign active discounts to `request.discounts`."""
 
     def middleware(request):
-        request.discounts = SimpleLazyObject(
-            lambda: fetch_discounts(datetime.date.today())
-        )
+        request.discounts = SimpleLazyObject(lambda: fetch_discounts(timezone.now()))
         return get_response(request)
 
     return middleware